+++ src/test/java/codechicken/nei/config/Option.java
@@ -1,0 +1,10 @@
+package codechicken.nei.config;
+
+public class Option {
+
+    public final String name;
+
+    protected Option(String name) {
+        this.name = name;
+    }
+}