+++ src/test/java/codechicken/nei/config/OptionKeyBind.java
@@ -1,0 +1,12 @@
+package codechicken.nei.config;
+
+public final class OptionKeyBind extends Option {
+
+    @SuppressWarnings("unused")
+    private final boolean useHash;
+
+    public OptionKeyBind(String identifier, boolean useHash) {
+        super("keys." + identifier);
+        this.useHash = useHash;
+    }
+}